split out pci and time related common utility functions order proto fields with optional last
