+++ src/control/server/mgmt_system_test.go
@@ -143,7 +143,6 @@ func TestServer_MgmtSvc_ClusterEvent(t *testing.T) {
 			expResp: &mgmtpb.ClusterEventResp{
 				Sequence: 1,
 			},
-			expErr: errors.New("unexpected event type"),
 		},
 		"unknown event type": {
 			rasEvent: &ras.Event{ID: math.MaxUint32},
